Logger and web UI used in first hatch
particle_logger.py polls the particle API for temperature/RH and logs
data to .csv files on disk. quail_maker_web.py is a Flask app that reads
the logged CSV data and presents it as a simple web page. It also allows
all data to be downloaded in a combines .csv file.